backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
cryptography==41.0.7
PyJWT==2.8.0
Flask-Mail==0.9.1
pyotp==2.9.0
qrcode==7.4.2
Pillow>=10.2.0
